refactor mutex to lock relevant part (periodStart); atomic int for counter; add test for sliding window

+++ handler/ratelimit/config_internal_test.go
@@ -18,9 +18,9 @@ func TestConfig_Errors(t *testing.T) {
 		min     string = "1m"
 		sec     string = "60s"
 		neg     string = "-1s"
-		num     uint   = 123
+		num     uint64 = 123
 		zeroStr string = "0s"
-		zeroInt uint   = 0
+		zeroInt uint64 = 0
 	)
 
 	for _, tc := range []testCase{
